megano/requirements.txt
asgiref==3.7.2
astroid==2.15.6
dill==0.3.7
diploma-frontend==0.6
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==10.0.0
platformdirs==3.10.0
pylint==2.17.5
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
wrapt==1.15.0
